+++ examples/http_client_usage/multiple_example.py
@@ -15,15 +15,13 @@ import pyjsonrpc
 rpc_client = pyjsonrpc.HttpClient("http://localhost:8080")
 
 # Example with *multiple calls* in one request
-methods = [
-    pyjsonrpc.create_request_dict("add", 1, 2),
-    pyjsonrpc.create_request_dict("add", 3, 4)
-]
-print rpc_client.call(methods)
+print rpc_client.call([
+    pyjsonrpc.Request.create("add", 1, 2),
+    pyjsonrpc.Request.create("add", 3, 4)
+])
 
 # Example with *multiple notifications* (no response) in one request
-notifications = [
-    pyjsonrpc.create_request_dict("add", 1, 2),
-    pyjsonrpc.create_request_dict("add", 3, 4)
-]
-rpc_client.notify(notifications)
+rpc_client.notify([
+    pyjsonrpc.Request.create("add", 1, 2),
+    pyjsonrpc.Request.create("add", 3, 4)
+])