+++ history.rst
@@ -8,6 +8,16 @@ by Gerold Penz 2013-2016
 - ToDo: Urllib3 for faster client requests
 
 
+=============
+Version 0.9.3
+=============
+
+2016-07-27
+
+- New method *Request.create*
+
+- New function *rpcrequest.create_request()*
+
 
 =============
 Version 0.9.2